+++ assets/js/phoenix_live_view/dom.js
@@ -4,7 +4,6 @@ import {
   DEBOUNCE_TRIGGER,
   FOCUSABLE_INPUTS,
   PHX_COMPONENT,
-  PHX_EVENT_CLASSES,
   PHX_HAS_FOCUSED,
   PHX_HAS_SUBMITTED,
   PHX_MAIN,
@@ -236,7 +235,9 @@ let DOM = {
 
       case "blur":
         if(this.once(el, "debounce-blur")){
-          el.addEventListener("blur", () => callback())
+          el.addEventListener("blur", () => {
+            if(asyncFilter()){ callback() }
+          })
         }
         return
 
@@ -446,16 +447,6 @@ let DOM = {
 
   isNowTriggerFormExternal(el, phxTriggerExternal){
     return el.getAttribute && el.getAttribute(phxTriggerExternal) !== null
-  },
-
-  syncPendingRef(fromEl, toEl, disableWith){
-    let ref = fromEl.getAttribute(PHX_REF)
-    if(ref === null){ return true }
-    let refSrc = fromEl.getAttribute(PHX_REF_SRC)
-
-    if(DOM.isUploadInput(fromEl)){ DOM.mergeAttrs(fromEl, toEl, {isIgnored: true}) }
-    DOM.putPrivate(fromEl, PHX_REF, toEl)
-    return false
   },
 
   cleanChildNodes(container, phxUpdate){