+++ CHANGELOG.md
@@ -108,17 +108,41 @@ generated user module:
      if valid_password?(changeset.data, password) do
 ```
 
-## 1.0.0-rc.1
+## 1.0.0-rc.3 (2024-06-20)
+
+### Bug fixes
+  * Fix issue where locked data-phx-ref containers would still be patched in some cases
+  * Fix issue where locked data-phx-ref containers would fail to have pending updates like streams applied after acknowledgment
+  * Fix issue where a LiveComponent issuing live navigation sends an acknowledgement too early before its parent LiveView handle_params diff is sent
+  * Fix issue allowing phx-remove elements to remain interactive while transitioning out from the page
+
+## 1.0.0-rc.2 (2024-06-19)
+
+### Bug fixes
+  * Fix bug introduced in rc.1 causing incorrect patching on form elements when awaiting acknowledgements
+
+## 1.0.0-rc.1 (2024-06-18)
 
 ### Enhancements
   * Warn when rendering an input named "id"
+  * Allow form recovery to work on forms in nested LiveViews
+  * Allow using form/3 with inputs outside the form
+  * Allow setting page title to an empty string
+  * Fix warnings on Elixir 1.17
 
 ### Bug fixes
   * Fix used input tracking on checkboxes and hidden inputs
+  * Fix phx-debounce=blur incorrectly sending change event to the next page in some cirumstances
+  * Fix race condition when destroying Live Components while transitions are still running
+  * Fix page reload when disconnecting LiveSocket if using Bandit
+  * Fix formatter changing `<%` to `<%=` when it shouldn't
 
 ### Deprecations
   * Deprecate `live_flash` in favor of `Phoenix.Flash.get`
   * Deprecate `push_redirect` in favor of `push_navigate`
+
+### Removal of previously deprecated functionality
+  * `phx-capture-click` has been removed (deprecated in 0.17.0)
 
 ## 1.0.0-rc.0 🚀 (2024-05-08)
 
@@ -133,9 +157,41 @@ generated user module:
   * Fix nested LiveView within streams becoming empty when reset
   * Fix `phx-mounted` firing twice, first on dead render, then on live render, leading to errors when a LiveComponent has not yet mounted
   * Fix `JS.toggle_class` error when used with a transition
+  * Fix updated form values not being sent in some circumstances when phx-trigger-action is used
+  * Fix phx-viewport bindings when html or body element has overflow property
 
 ### Enhancements
   * Warn on mismatched client and server versions
+
+## 0.20.15 (2024-06-18)
+
+### Enhancements
+  * Warn when rendering an input named "id"
+  * Warn on mismatched client and server versions
+  * Allow form recovery to work on forms in nested LiveViews
+  * Allow using form/3 with inputs outside the form
+  * Allow setting page title to an empty string
+  * Fix warnings on Elixir 1.17
+
+### Bug fixes
+  * Fix attributes of existing stream items not being updated on reset
+  * Fix nested LiveView within streams becoming empty when reset
+  * Fix `phx-mounted` firing twice, first on dead render, then on live render, leading to errors when a LiveComponent has not yet mounted
+  * Fix `JS.toggle_class` error when used with a transition
+  * Fix phx-debounce=blur incorrectly sending change event to the next page in some cirumstances
+  * Fix race condition when destroying Live Components while transitions are still running
+  * Fix page reload when disconnecting LiveSocket if using Bandit
+  * Fix formatter changing `<%` to `<%=` when it shouldn't
+  * Fix updated form values not being sent in some circumstances when phx-trigger-action is used
+  * Fix phx-viewport bindings when html or body element has overflow property
+
+### Deprecations
+  * Deprecate `live_flash` in favor of `Phoenix.Flash.get`
+  * Deprecate `push_redirect` in favor of `push_navigate`
+
+### Removal of previously deprecated functionality
+  * `phx-capture-click` has been removed (deprecated in 0.17.0)
+  * `live_component/2` and `live_component/3` helpers (not the function component) have been removed
 
 ## 0.20.14 (2024-03-13)
 
@@ -414,7 +470,7 @@ generated user module:
 ### Enhancements
   * Support [`submitter`](https://developer.mozilla.org/en-US/docs/Web/API/SubmitEvent/submitter) on form submit events.
   * Avoid compile-time dependency for `attr` when referencing structs
-  * Validate reserved assigns. Attempting to assign `:uploads`, `:streams`, `:live_action`, `:socket`, `:myself` will now raise in `LiveView` and `LiveComponent`
+  * Validate reserved assigns. Attempting to assign `:uploads`, `:streams`, `:live_action`, `:socket`, `:myself` will now raise in LiveView and LiveComponent
 
 ## 0.18.16 (2023-02-23)
 
@@ -1469,7 +1525,7 @@ The new implementation will check there is a button at `#term .buttons a`, with 
   - Add `put_live_layout` plug to put the root layout used for live routes
   - Allow `redirect` and `push_redirect` from mount
   - Use acknowledgement tracking to avoid patching inputs until the server has processed the form event
-  - Add css loading states to all phx bound elements with event specific css classes
+  - Add CSS loading states to all phx bound elements with event specific CSS classes
   - Dispatch `phx:page-loading-start` and `phx:page-loading-stop` on window for live navigation, initial page loads, and form submits, for user controlled page loading integration
   - Allow any phx bound element to specify `phx-page-loading` to dispatch loading events above when the event is pushed
   - Add client side latency simulator with new `enableLatencySim(milliseconds)` and `disableLatencySim()`
@@ -1675,12 +1731,12 @@ let liveSocket = new LiveSocket("/live", Socket, {...})
 
 ### Backwards incompatible changes
   - `phx-value` has no effect, use `phx-value-*` instead
-  - The `:path_params` key in session has no effect (use `handle_params` in `LiveView` instead)
+  - The `:path_params` key in session has no effect (use `handle_params` in LiveView instead)
 
 ## 0.1.1 (2019-08-27)
 
 ### Enhancements
-  - Use optimized `insertAdjacentHTML` for faster append/prepend and proper css animation handling
+  - Use optimized `insertAdjacentHTML` for faster append/prepend and proper CSS animation handling
   - Allow for replacing previously appended/prepended elements by replacing duplicate IDs during append/prepend instead of adding new DOM nodes
 
 ### Bug Fixes